changed project sturcture
