+++ models/auth.go
@@ -1,0 +1,24 @@
+package models
+
+import (
+	"time"
+
+	"github.com/golang-jwt/jwt/v5"
+)
+
+type Tokens struct {
+	AccessToken  string `json:"access_token"`
+	RefreshToken string `json:"refresh_token"`
+}
+
+type User struct {
+	Username  string    `json:"username"`
+	CreatedAt time.Time `json:"created_at"`
+	UpdatedAt time.Time `json:"updated_at"`
+	Password  string    `json:"password"`
+}
+
+type Claims struct {
+	User User `json:"user"`
+	jwt.RegisteredClaims
+}
