+++ models/cash.go
@@ -1,0 +1,25 @@
+package models
+
+import (
+	"time"
+
+	"github.com/google/uuid"
+)
+
+type CashBody struct {
+	APIKey  string  `json:"api_key" binding:"required"`
+	Amount  float64 `json:"amount" binding:"required"`
+	Contact string  `json:"contact" binding:"required"`
+	Detail  string  `json:"detail"`
+	Note    string  `json:"note"`
+}
+
+type CashBodyResponse struct {
+	UUID      uuid.UUID `json:"uuid"`
+	Amount    float64   `json:"amount" binding:"required"`
+	Detail    string    `json:"detail"`
+	Note      string    `json:"note"`
+	Client    string    `json:"client"`
+	Contact   string    `json:"contact"`
+	CreatedAt time.Time `json:"created_at"`
+}
